+++ src/frontend/src/lib/components/ui/EmptyState.svelte
@@ -1,10 +1,11 @@
 <script lang="ts">
+	import { notEmptyString } from '@dfinity/utils';
 	import { fade } from 'svelte/transition';
 	import IconOisyEmptyState from '$lib/components/icons/IconOisyEmptyState.svelte';
 
 	interface Props {
 		title: string;
-		description: string;
+		description?: string;
 	}
 
 	let { title, description }: Props = $props();
@@ -13,5 +14,8 @@
 <div in:fade class="my-10 flex flex-col items-center justify-center px-4 text-center">
 	<IconOisyEmptyState />
 	<div class="mb-1 mt-5 font-bold">{title}</div>
-	<div class="text-sm">{description}</div>
+
+	{#if notEmptyString(description)}
+		<div class="text-sm">{description}</div>
+	{/if}
 </div>
